pyproject.toml
[tool.poetry]
name = "tweet"
version = "1.0.0"
description = ""
authors = ["ganariya <ganariya2525@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/Ganariya/tweet"

[tool.poetry.scripts]
tweet = "tweet.main:app"

[tool.poetry.dependencies]
python = "^3.9"
typer = { extras = ["all"], version = "^0.3.2" }
black = "^20.8b1"
mypy = "^0.790"
flake8 = "^3.8.4"
python-dotenv = "^0.15.0"
python-twitter = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
